extract keyboard shortcuts
